+++ conjurapi/policy_test.go
@@ -2,12 +2,13 @@ package conjurapi
 
 import (
 	"fmt"
-	"github.com/cyberark/conjur-api-go/conjurapi/authn"
-	"github.com/cyberark/conjur-api-go/conjurapi/response"
-	. "github.com/smartystreets/goconvey/convey"
 	"os"
 	"strings"
 	"testing"
+
+	"github.com/cyberark/conjur-api-go/conjurapi/authn"
+	"github.com/cyberark/conjur-api-go/conjurapi/response"
+	. "github.com/smartystreets/goconvey/convey"
 )
 
 func TestClient_LoadPolicy(t *testing.T) {
@@ -15,19 +16,20 @@ func TestClient_LoadPolicy(t *testing.T) {
 		config := &Config{}
 		config.mergeEnv()
 
-		api_key := os.Getenv("CONJUR_AUTHN_API_KEY")
+		apiKey := os.Getenv("CONJUR_AUTHN_API_KEY")
 		login := os.Getenv("CONJUR_AUTHN_LOGIN")
 
 		Convey("Successfully load policy", func() {
-			variable_identifier := "alice"
+			variableIdentifier := "alice"
 			policy := fmt.Sprintf(`
 - !user %s
-`, variable_identifier)
+`, variableIdentifier)
 
-			conjur, err := NewClientFromKey(*config, authn.LoginPair{login, api_key})
+			conjur, err := NewClientFromKey(*config, authn.LoginPair{login, apiKey})
 			So(err, ShouldBeNil)
 
 			resp, err := conjur.LoadPolicy(
+				PolicyModePut,
 				"root",
 				strings.NewReader(policy),
 			)
@@ -40,10 +42,10 @@ func TestClient_LoadPolicy(t *testing.T) {
 			login = "invalid-user"
 
 			Convey("Returns 401", func() {
-				conjur, err := NewClientFromKey(*config, authn.LoginPair{login, api_key})
+				conjur, err := NewClientFromKey(*config, authn.LoginPair{login, apiKey})
 				So(err, ShouldBeNil)
 
-				resp, err := conjur.LoadPolicy("root", strings.NewReader(""))
+				resp, err := conjur.LoadPolicy(PolicyModePut, "root", strings.NewReader(""))
 
 				So(err, ShouldNotBeNil)
 				So(resp, ShouldBeNil)
@@ -63,18 +65,19 @@ func TestClient_LoadPolicy(t *testing.T) {
 		}
 
 		login := os.Getenv("CONJUR_V4_AUTHN_LOGIN")
-		api_key := os.Getenv("CONJUR_V4_AUTHN_API_KEY")
+		apiKey := os.Getenv("CONJUR_V4_AUTHN_API_KEY")
 
 		Convey("Policy loading is not supported", func() {
-			variable_identifier := "alice"
+			variableIdentifier := "alice"
 			policy := fmt.Sprintf(`
 - !user %s
-`, variable_identifier)
+`, variableIdentifier)
 
-			conjur, err := NewClientFromKey(*config, authn.LoginPair{login, api_key})
+			conjur, err := NewClientFromKey(*config, authn.LoginPair{login, apiKey})
 			So(err, ShouldBeNil)
 
 			_, err = conjur.LoadPolicy(
+				PolicyModePut,
 				"root",
 				strings.NewReader(policy),
 			)