+++ conjurapi/router_v4.go
@@ -14,10 +14,10 @@ type RouterV4 struct {
 	Config *Config
 }
 
-func (self RouterV4) AuthenticateRequest(loginPair authn.LoginPair) (*http.Request, error) {
-	authenticateUrl := fmt.Sprintf("%s/authn/users/%s/authenticate", self.Config.ApplianceURL, url.QueryEscape(loginPair.Login))
+func (r RouterV4) AuthenticateRequest(loginPair authn.LoginPair) (*http.Request, error) {
+	authenticateURL := fmt.Sprintf("%s/authn/users/%s/authenticate", r.Config.ApplianceURL, url.QueryEscape(loginPair.Login))
 
-	req, err := http.NewRequest("POST", authenticateUrl, strings.NewReader(loginPair.APIKey))
+	req, err := http.NewRequest("POST", authenticateURL, strings.NewReader(loginPair.APIKey))
 	if err != nil {
 		return nil, err
 	}
@@ -26,13 +26,13 @@ func (self RouterV4) AuthenticateRequest(loginPair authn.LoginPair) (*http.Reque
 	return req, nil
 }
 
-func (self RouterV4) RotateAPIKeyRequest(roleId string) (*http.Request, error) {
-	tokens := strings.SplitN(roleId, ":", 3)
+func (r RouterV4) RotateAPIKeyRequest(roleID string) (*http.Request, error) {
+	tokens := strings.SplitN(roleID, ":", 3)
 	if len(tokens) != 3 {
-		return nil, fmt.Errorf("Role id '%s' must be fully qualified", roleId)
+		return nil, fmt.Errorf("Role id '%s' must be fully qualified", roleID)
 	}
-	if tokens[0] != self.Config.Account {
-		return nil, fmt.Errorf("Account of '%s' must match the configured account '%s'", roleId, self.Config.Account)
+	if tokens[0] != r.Config.Account {
+		return nil, fmt.Errorf("Account of '%s' must match the configured account '%s'", roleID, r.Config.Account)
 	}
 
 	var username string
@@ -43,45 +43,45 @@ func (self RouterV4) RotateAPIKeyRequest(roleId string) (*http.Request, error) {
 		username = strings.Join([]string{tokens[1], tokens[2]}, "/")
 	}
 
-	rotateUrl := fmt.Sprintf("%s/authn/users/api_key?id=%s", self.Config.ApplianceURL, username)
+	rotateURL := fmt.Sprintf("%s/authn/users/api_key?id=%s", r.Config.ApplianceURL, username)
 
 	return http.NewRequest(
 		"PUT",
-		rotateUrl,
+		rotateURL,
 		nil,
 	)
 }
 
-func (self RouterV4) LoadPolicyRequest(policyId string, policy io.Reader) (*http.Request, error) {
+func (r RouterV4) LoadPolicyRequest(mode PolicyMode, policyID string, policy io.Reader) (*http.Request, error) {
 	return nil, fmt.Errorf("LoadPolicy is not supported for Conjur V4")
 }
 
-func (self RouterV4) CheckPermissionRequest(resourceId, privilege string) (*http.Request, error) {
-	tokens := strings.SplitN(resourceId, ":", 3)
+func (r RouterV4) CheckPermissionRequest(resourceID, privilege string) (*http.Request, error) {
+	tokens := strings.SplitN(resourceID, ":", 3)
 	if len(tokens) != 3 {
-		return nil, fmt.Errorf("Resource id '%s' must be fully qualified", resourceId)
+		return nil, fmt.Errorf("Resource id '%s' must be fully qualified", resourceID)
 	}
-	checkUrl := fmt.Sprintf("%s/authz/%s/resources/%s/%s?check=true&privilege=%s", self.Config.ApplianceURL, tokens[0], tokens[1], url.QueryEscape(tokens[2]), url.QueryEscape(privilege))
+	checkURL := fmt.Sprintf("%s/authz/%s/resources/%s/%s?check=true&privilege=%s", r.Config.ApplianceURL, tokens[0], tokens[1], url.QueryEscape(tokens[2]), url.QueryEscape(privilege))
 
 	return http.NewRequest(
 		"GET",
-		checkUrl,
+		checkURL,
 		nil,
 	)
 }
 
-func (self RouterV4) AddSecretRequest(variableId, secretValue string) (*http.Request, error) {
+func (r RouterV4) AddSecretRequest(variableIDentifier, secretValue string) (*http.Request, error) {
 	return nil, fmt.Errorf("AddSecret is not supported for Conjur V4")
 }
 
-func (self RouterV4) RetrieveSecretRequest(variableIdentifier string) (*http.Request, error) {
+func (r RouterV4) RetrieveSecretRequest(variableIDentifier string) (*http.Request, error) {
 	return http.NewRequest(
 		"GET",
-		self.variableURL(variableIdentifier),
+		r.variableURL(variableIDentifier),
 		nil,
 	)
 }
 
-func (self RouterV4) variableURL(variableIdentifier string) string {
-	return fmt.Sprintf("%s/variables/%s/value", self.Config.ApplianceURL, url.QueryEscape(variableIdentifier))
+func (r RouterV4) variableURL(variableIDentifier string) string {
+	return fmt.Sprintf("%s/variables/%s/value", r.Config.ApplianceURL, url.QueryEscape(variableIDentifier))
 }