+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # v0.3.0
 
 * Adds new API methods `RotateAPIKey` and `CheckPermission`.
-* Methods that return secret data return an `io.ReadCloser` instead of `[]byte`. This way, the API client gets the only copy of the secret data and can handle it however she sees fit.
+* Provide API methods that return secret data as an `io.ReadCloser` rather than of `[]byte`. This way, the API client gets the only copy of the secret data and can handle it however she sees fit.
+* Loading a policy requires `PolicyMode` argument.
 
 # v0.2.0